requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
dj-database-url==2.2.0
Django==4.2.30
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
gunicorn==23.0.0
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==26.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
PyYAML==6.0.3
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.5
typing_extensions==4.13.2
uritemplate==4.1.1
whitenoise==6.7.0
zipp==3.20.2
